example-lambda/utils/hcl-to-env-vars/requirements.txt
ply >= 3.11
pyhcl >= 0.3.12
